requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
llama-index-embeddings-gemini
streamlit
google-api-python-client
llama-index-llms-gemini
llama-index-readers-web
html2text
llama-index-vector-stores-postgres